+++ Code/Mantid/Framework/API/src/TableRow.cpp
@@ -14,7 +14,7 @@ Kernel::Logger& TableRow::g_log = Kernel::Logger::get("TableRow");
   */
 TableRow::TableRow(const TableRowHelper& trh):m_row(trh.m_row),m_col(0),m_sep(",")
 {
-    for(int i=0;i<trh.m_workspace->columnCount();i++)
+    for(size_t i=0;i<trh.m_workspace->columnCount();i++)
         m_columns.push_back(trh.m_workspace->getColumn(i));
     if (m_columns.size()) m_nrows = int(m_columns[0]->size());
     else
