+++ Code/Mantid/Framework/Algorithms/src/StripPeaks.cpp
@@ -153,7 +153,7 @@ API::MatrixWorkspace_sptr StripPeaks::removePeaks(API::MatrixWorkspace_const_spt
   //progress from 0.3 to 1.0 here 
   prg=0.3;
   // Loop over the list of peaks
-  for (int i = 0; i < peakslist->rowCount(); ++i)
+  for (size_t i = 0; i < peakslist->rowCount(); ++i)
   {
     // Get references to the data
     const MantidVec &X = outputWS->readX(peakslist->getRef<int>("spectrum",i));
