+++ Code/Mantid/MantidQt/MantidWidgets/src/ICatUtils.cpp
@@ -45,21 +45,21 @@ namespace MantidQt
       tablewidget->setRowCount(ws_sptr->rowCount());
       tablewidget->setColumnCount(ws_sptr->columnCount());
 
-      for (int i=0;i<ws_sptr->rowCount();++i)
+      for (size_t i=0;i<ws_sptr->rowCount();++i)
       {
         //setting the row height of tableWidget
         tablewidget->setRowHeight(i,20);
       }
 
       QStringList qlabelList;
-      for(int i=0;i<ws_sptr->columnCount();i++)
+      for(size_t i=0;i<ws_sptr->columnCount();i++)
       {
         Column_sptr col_sptr = ws_sptr->getColumn(i);
         //get the column name to display as the header of table widget
         QString colTitle = QString::fromStdString(col_sptr->name());
         qlabelList.push_back(colTitle);
 
-        for(int j=0;j<ws_sptr->rowCount();++j)
+        for(size_t j=0;j<ws_sptr->rowCount();++j)
         {
           std::ostringstream ostr;
           col_sptr->print(j,ostr);
